pytest_framework/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
debugpy==1.8.0
decorator==5.1.1
fastjsonschema==2.18.1
iniconfig==2.0.0
json-compare==0.2.11
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Pillow==10.1.0
Pillow-PIL==0.1.dev0
psutil==5.9.6
pure-eval==0.2.2
py4j==0.10.9.5
pytest==7.3.2
pytest-base-url==2.0.0
pytest-check==2.2.2
pytest-html==4.1.1
pytest-metadata==3.0.0
pytest-playwright==0.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-rapidjson==1.13
python-slugify==8.0.1
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
selenium==4.8.3
sqlparse==0.4.4
types-python-dateutil==2.8.19.14
typing_extensions==4.6.3
urllib3==1.26.15
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4

